requirements.txt
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.10
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
pypdf>=3.17.4
unstructured>=0.10.30
tiktoken>=0.5.1
